source-posthog/pyproject.toml
[tool.poetry]
name = "source-posthog"
version = "0.1.0"
description = "Estuary Flow source connector for PostHog analytics"
readme = "README.md"
authors = ["CorrDyn"]

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2"
python-dotenv = "^1.0"
requests = "^2.32"
estuary-cdk = { git = "https://github.com/estuary/connectors", subdirectory = "estuary-cdk" }

[tool.poetry.group.dev.dependencies]
debugpy = "^1.8"
mypy = "^1.8"
pytest = "^7.4"
pytest-asyncio = "^0.23"
pytest-insta = "^0.3"
ruff = "^0.8"
pre-commit = "^4.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "RUF",    # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
